Load actual day name, tidied up card styling and code
