+++ App.js
@@ -1,7 +1,6 @@
-import { useState, useEffect } from 'react'
-import { StatusBar } from 'expo-status-bar';
+import { useState, useEffect } from 'react';
 import { StyleSheet, Text, View, Button, TextInput } from 'react-native';
-import Page1 from './src/pages/Page1';
+import DailyWeatherCards from './src/pages/DailyWeatherCards';
 import { GetWeatherForecast } from './src/api/api';
 
 export default function App() {
@@ -14,25 +13,19 @@ export default function App() {
 
   const HandleSearchButton = async () => {
     var result = await GetWeatherForecast(userInput);
-    setWeeklyForecast(result); 
+    setWeeklyForecast(result);
   }
 
   return (
     <View style={styles.container}>
-      <Text>{userInput}</Text>
-      <StatusBar style="auto" />
-      <View style={styles.cardSection}>
-        {/* <Page1 />
-        <Page1 />
-        <Page1 />
-        <Page1 />
-        <Page1 />
-        <Page1 />
-        <Page1 /> */}
-      </View>
       <View style={styles.cardSection}>
         {weeklyForecast.map((data) => (
-          <Page1 day={data.date} dailyMaxTemp={data.tempMax} dailyMinTemp={data.tempMin}/>
+          <DailyWeatherCards
+            key={data.date}
+            day={data.date}
+            dailyMaxTemp={data.tempMax}
+            dailyMinTemp={data.tempMin}
+          />
         ))}
       </View>
       {/* <TextInput
@@ -49,9 +42,9 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: '#fff',
-    alignItems: 'center',   
+    alignItems: 'center',
     justifyContent: 'center',
-    // marginTop: 80,    
+    // marginTop: 80,
   },
   textInput: {
     height: 40,
@@ -65,5 +58,5 @@ const styles = StyleSheet.create({
     justifyContent: 'space-between',
     margin: 5,
     padding: 10
-  }  
+  }
 });
